backend-python/requirements.txt
google-cloud-storage
mysql-connector-python
flask
scikit-learn
numpy
